refactored everything out of App.js
